Adding some code to Login.js
I was working with the Login.js file and comitted a few changes that I made while trying to figure things out. I did not intentionally change the package-lock.json file, but I'm guessing that happened as a result of me making changes and running the app or installing software to run the app.